Improve event list page styling
