+++ frontend/src/interfaces/index.ts
@@ -10,3 +10,19 @@ export interface User {
   permissions: string[]
   groups?: Group[]
 }
+
+export interface APIResponseWithCollection<T> {
+  count: number;
+  next: number | null;
+  previous: number | null;
+  results: T[];
+}
+
+export interface Event {
+  id: number;
+  title: string;
+  description: string;
+  location: string;
+  start_time: string;
+  end_time: string;
+}
