+++ frontend/src/pages/Events/index.tsx
@@ -1,11 +1,16 @@
+import { Button, Container } from 'react-bootstrap';
+
 import { EventsList } from '../../components/EventList';
 
 export function Events() {
   return (
-    <div>
-      <h1>Events</h1>
+    <Container className="mt-4">
+      <div className="d-flex justify-content-between align-items-center mb-3">
+        <h1 className="mb-0">Events</h1>
+        <Button variant="primary">Create Event</Button>
+      </div>
 
-      <EventsList></EventsList>
-    </div>
+      <EventsList />
+    </Container>
   );
 }
