+++ frontend/src/utils/formatDate.ts
@@ -1,0 +1,6 @@
+export function formatDate(iso: string): string {
+  return new Date(iso).toLocaleString(undefined, {
+    dateStyle: 'medium',
+    timeStyle: 'short',
+  });
+}
